v0.0.2 Player Movement
The player's movement was altered to be calculated in a different way, altering its velocity instead of its position.

The player's base speed was reduced to fit better the new way of calculating movement.

The walking and running animations were given exit time so that they don't get interrupted. This should solve the issue of the player going through the idle animation when suddenly changing direction while moving.