v0.1.0 Lantern
The basic functionality of the lantern is now implemented. The player can summon or unsummon it using Right Click. During the summoning animation, the player can not move.

The lamp rotates around the player when summoned, following the mouse.

The lamp was given a slight flicker animation.

The player was given a provisional Jump animation when pressing space. Jumping fully is yet to be implemented.

A "Player" tag was created for later use.