v0.3.0 Lantern Light
The lantern is now a prefab.

The lantern now illuminates an area in front of it, made from a triangle and a circle.

The Lantern Summon key was changed from Right Click to "E".

The player can switch between colors using Right Click and return to the default color using Left Click.