v0.4.0 Color Swap Rework
The system used to swap Lantern colors was reworked. It now uses a list of colors that can be set through the inspector. The part of the script for this was adjusted accordingly so that it can read from the list. The name of the method was also changed to reference the lantern.

A new method was created so that the player switches its sprite's color to that of the lantern's light (of course, only when the lantern is active - otherwise it returns to the default color when invoked).

A "Sit" animation was added for the player. During this short animation, the mentioned method to change the player's color is invoked. This animation was given the "S" key.

The Player prefab was updated to reflect these changes.

This time, the scene was also saved before committing.

Several comments were added to PlayerController.cs for a better understanding.