v0.6.0 Jump
The player can now jump. A ground check child object was given to the player. Several variables related to this were added in PlayerController.cs.

The player Prefab was updated to reflect the previous changes. It was also given more gravity scale and drag in order to not be sent flying when going up some stairs.

A new layer for "Ground" was added.

This method of implementing ground check could be considered a crime in 2D Design class, since it uses a child GameObject rather than a Raycast. Proceed with caution.