v0.6.1 Player robe color
The player's robe was changed from red to white to better reflect the changes in color when using the lamp.

The number of unlocked colors by default has been updated to 5 (max) in order to simplify the testing process. This will be reverted once it is needed.

The ground check child object was moved slightly downwards and its check area has been increased.