v0.7.0 Double jump
The player can now double jump. Optionally, the player can be given a higher number of max jumps.

The jump is now handled by its own method in PlayerController.cs. This method checks the player's speed in the air and determines the multiplier to apply to the force added to jump.

Several variables were created for this purpose.

The player prefab was updated.