v0.8.3 Double jump animation
When double / extra jumping, the animation will start once again. This was made by making a copy of the jump animation and having both transition between each other when jump is pressed.

The Animator trigger for jumping was moved from HandleAnimations method to HandleJump method, since it would require more checks for it to be implemented there.