v0.9.1 Ground check fix
The ground check box was reduced so that it no longer detects walls the player is close to. Consequently, coyote time is no longer available.

The jump animation(s) can now transition to the summoning lantern animation. This means the jump animation will get interrupted.

The player prefab was adjusted accordingly.