v0.10.0 Life and Death
The player can now receive damage from enemies. The method was updated so that it gets the damage variable set for each enemy to deal to the player.

The previous knockback method was removed. Currently, only death applies knockback and frankly, I don't even know why.

The player can now die and it will trigger a death animation. Shortly after the death animation, a revival method will be called.

In the revival method, the user can choose the first SpawnPoint the player will appear at (a temporary one was created in the scene).

The player prefab was updated with all needed changes.

A temporary enemy (staticsphere) was added, along with a script that can deal damage to the player. An "Enemy" tag was also created.