v0.11.4 Minor changes
The player no longer starts with a forced 1 color. The code responsible has been commented and it can now be set through the inspector.

The player prefab was updated so that doesn't include variables that are no longer public.

The lantern was given a check to avoid out of bounds errors when the Unlocked Colors is set to 0 or less.