v0.12.0 Color Benefits
The Player's color now determines some benefits, like green givning more speed or yellow giving more jump force and an additional jump. White does not do anything. In the future, when damage is implemented, it is intended for the player to deal more damage while being red. Blue is still to be determined.

The PlayerController.cs script has seen some changes in variables due to this change. Additional variables to remember the previous variable states (only variables that change based on color) before the color swap.

The player prefab was adjusted accordingly.

The Level1 Scene has been slightly modified to be more immersive and intuitive to get through.