v0.13.5 Background follow camera
The background images in level 2 now follows the Main Camera. The Main Camera object must be bassed to the background parent through the inspector, along with the vertical offset.

A prefab was made for the Background parent.

Some sections of level 1 were adjusted so the player can actually progress.

Scripts related to the camera or background were moved to the "Camera" folder.