v0.14.0 Fall death
A new object was made: Deathbox. It's a trigger collider that kills the player on contact. It is currently used to detect falls. Level 1 and Level 2 were each given one of these colliders.

A new tag was added for the player to die when touching one of these colliders.

The CameraController.cs script was given an extra parameter that checks whether the player's position is under said variable. If so, it stops following in the y axis. Useful when the player falls for the camera to not go along with it.

The part of the PlayerController.cs script that handles death was moved to its own method for it to work properly. In the future, this method may be used to control other player life parameters.