v0.16.0 Player Health UI
A new script was created to handle player Health Flames. All the flames must be passed twice through the Inspector, one for the gameobject and another for the animator.

The player health now shows accurately in the UI (up to a maximum of 5).

The animations related to the flames play properly.

The player prefab was adjusted to reflect these changes.