v0.17.7 Color Swap rework
Now the Player swaps color to the lamp color without needing to sit. The color will reset only when sitting without the lamp or when changing to the default color.

The player can now have different actions when pressing right click based on the color. There is a cooldown to these actions.

The Player prefab was adjusted accordingly.