v0.18.1 Lantern action expansion
The lantern actions have been slightly updated so that it now gets activated during a new animation, which is just the summoning animation but faster.

The enemies no longer respawn and in the future it will just reload the scene.

The level 1 tilemap was changed so that the level is more coherent in design and tilemap.

Jumping while in inverted gravity has been fixed.

A new tag was added so that the light from the lantern doesn't collide with the enemies. However, this now means that the light won't deal damage, which will be fixed in the future.

The player and lantern prefabs were adjusted accordingly.