v0.21.0 Boss Ghost Implemented
A boss Ghost was implemented. It starts idle. Should the player get close, it would play an animation and start moving. It alternates between following the player and going between different patrol points.

A prefab was made for the new enemy.

Several files were reorganised into folders, such as prefabs.

A level 1 skeleton was adjusted so that it is now a prefab rather than a prefab variant.

The skeleton from the sample scene was removed in order to give space to the new boss enemy.