v0.21.2 Boss ghost flip
The Boss Ghost's sprite can now flip. Previously, it wasn't able to because of Unity Shenanigans not allowing it to transform the localscale during an animation.

This was solved using a pivot parent object that contains the script and a child gameobject that contains everything else.

Several prefabs were adjusted, as well as most scripts involved.