v0.21.4 BossGhost HItbox
The Boss Ghost's hitbox was adjusted to be more lenient and not damage the player when it appears that it shouldn't.

The method that handles life in the PlayerController.cs script was moved outside the conditional that checks if the player is dead because it apparently causes errors when the player's hp is reduced to 0.