v0.24.1 Updated Light Colors
The lantern's light now works properly and has a larger collider to correspond with the new way of rendering light.

The PlayerController.cs script was adjusted so that it now swaps the color of the light according to the color of the lamp. Rather than copying its color, it sets the color to the maximum value of red, blue, green or yellow.

Although it doesn't appear as a change, the following line was added to the beginning of the script in order for the light to work:
"using UnityEngine.Experimental.Rendering.Universal;"

Levels 2 and 3 were adjusted to include the global light.

The prefabs for the Lantern and Player were adjusted.