v0.25.0 Color scroll wheel
The player now switches its color using the mouse scroll wheel. The alternate color key was also changed from Q to right click, since due to the change, it was then empty.

Several prefabs were adjusted and a canvas prefab was created in order to change the UI in all levels at once.