v0.26.10 Color swap sit
The player's color can now be changed while the lamp is not summoned by sitting. This has no effect while the lamp is summoned because it already sets the player to the lamp's color.

The skeleton's hitbox was adjusted so that it fits the sprite better.

Some ghosts were added to level 1.