v0.28.0 Next & Previous Colors
There are now 2 more color frames that display the previous and the next color.

New tags were added for the player to find the new objects without setting them through the inspector.

Everything in the player's UI has been moved to the top left corner.

The player in level 2 was moved from its testing position back to its original position.