v0.28.3 Boss Fight trigger
The Boss Fight in level 2 now only starts once the player enters the room. Until then, the boss is idle.

A force field was added that is activated once the player enters the  room. As of now, the force field doesn't disappear.

The Force field was given the same layer as death pits because otherwise the player can cheese the boss fight by triggering its "scare" animation multiple times. This is just to not create a new layer.

The Boss Ghost has had its hitbox reduced.

The Boss Ghost prefabs were updated accordingly.