v0.31.2 Wizard Phase 2
The wizard now transitions into phase 2 at half health. In this phase, its attacks are more frequent, meaning it shoots more proyectiles. In addition, it now teleports to the next teleport point in less hits.

A new animation was made for the phase transition.

A slight glow was added to fireballs.