v0.32.0 Shadow Wizard Boss
A new wizard was added. It is mostly based on the previous one, but with some significant changes.

The boss can invert its own gravity while in Phase 2 and is more aggresive towards the player when provoked. However, as of now, it won't attack the player unless it approaches or attacks first.

Several animations were made, as well as a new spritefor the shadow ball proyectile (which is just a recolor of the fireball).

The wizard was added in the sample scene for testing.