v0.36.0 Level transitions
There are now box colliders at the end of levels which when in contact with the player, they transport it to the next level.

The collider only gets activated during the death animation of bosses.

A white fading animation was made for when the player transitions to a new level.